FEATURE: Support for asset variant presets
Provides a new property `variantPreset` in the `AssetImageSource` and
`DummyImageSource` Fusion prototypes. The expect a string giving the
preset identifier and preset variant name separated by `::`, like:

    Acme.Com:Component.Teaser.Hero::landscape

In addition this adds a new method `useVariantPreset()` to the
`ImageSourceHelperInterface` that can be used to specify a variant to
use based on a configured asset variant preset:

    useVariantPreset('Acme.Com:Component.Teaser.Hero', 'landscape')

That variant (if existing) of an asset will be used as base before any
further adjustments are applied.